Nile University's Arabic sentiment Lexicon NilULex v0.27
This file contains Egyptian Arabic and Modern Standard Arabic sentiment words and their polarity. This lexicon can be used freely for research purposes. For commercial use, you need to obtain written approval from the author: Samhaa R. El-Beltagy. The lexicon is provided as is, with no warranty. The appearance of any of the terms provided in this lexicon within a sentence does not necessarily indicate a corresponding sentiment.
If you use this lexicon, please cite the following paper: El-Beltagy, Samhaa R., 2016. NileULex: A Phrase and Word Level Sentiment Lexicon for Egyptian and Modern Standard Arabic. In proceedings of LREC 2016,Portorož , Slovenia. The paper can be found in this repository